refactor(put all the css modules file content into template.ts): css modules files refactored
put all the css modules file content into template.ts
