--- packages/cli/src/commands/styles/lib/templates/css-modules/elevation.ts
@@ -1,11 +1,0 @@
-import { RenderTokenFile, TOKEN_FILE_COMMENTS } from '../types';
-import { generateElevationCSS } from '../../../utils/cssGenerator.utils';
-
-export const elevation: RenderTokenFile = (tokens, type) =>
-  [
-    `./styles/_${ type }.css`,
-    `:root {
-${ TOKEN_FILE_COMMENTS[type] }
-${ tokens.map(( token ) => `${ generateElevationCSS(token) }`).join('\n') }
-}`
-  ] as const;
